+++ src/framework/framework_application.js
@@ -411,8 +411,12 @@ pc.extend(pc, function () {
         _parseApplicationProperties: function (props, callback) {
             this._width = props['width'];
             this._height = props['height'];
+            if (props['use_device_pixel_ratio']) {
+                this.graphicsDevice.maxPixelRatio = window.devicePixelRatio;
+            }
+
             this.setCanvasResolution(props['resolution_mode'], this._width, this._height);
-            this.setCanvasFillMode(props['fill_mode'], this._width, this._height)
+            this.setCanvasFillMode(props['fill_mode'], this._width, this._height);
 
             this._loadLibraries(props['libraries'], callback);
         },